@LukaszRozmej 2022-09-26
We should also try to get the sender from the Receipt where it should have already been recovered before going to ecdsa, we store receipts per block so it's just one hit to DB per call.